Validate entry.Name chars before extracting
