+++ src/libraries/Common/src/System/IO/Archiving.Utils.Unix.cs
@@ -1,0 +1,10 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+namespace System.IO
+{
+    internal static partial class ArchivingUtils
+    {
+        internal static string SanitizeEntryFilePath(string entryPath) => entryPath.Replace('\0', '_');
+    }
+}
