+++ src/libraries/Common/src/System/IO/Archiving.Utils.Windows.cs
@@ -1,17 +1,16 @@
-﻿// Licensed to the .NET Foundation under one or more agreements.
+// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
-
 
 using System.Text;
 
-namespace System.IO.Compression
+namespace System.IO
 {
-    public static partial class ZipFileExtensions
+    internal static partial class ArchivingUtils
     {
-        internal static string SanitizeZipFilePath(string zipPath)
+        internal static string SanitizeEntryFilePath(string entryPath)
         {
-            StringBuilder builder = new StringBuilder(zipPath);
-            for (int i = 0; i < zipPath.Length; i++)
+            StringBuilder builder = new StringBuilder(entryPath);
+            for (int i = 0; i < entryPath.Length; i++)
             {
                 if (((int)builder[i] >= 0 && (int)builder[i] < 32) ||
                    builder[i] == '?' || builder[i] == ':' ||
