+++ src/libraries/System.Formats.Tar/src/System/Formats/Tar/TarFile.cs
@@ -4,7 +4,6 @@
 using System.Buffers;
 using System.Diagnostics;
 using System.IO;
-using System.IO.Compression;
 using System.Threading;
 using System.Threading.Tasks;
